match_score: Float -> Integer (Phase 3.8) (#17)
Closes the last audit item. match_score always holds a rounded 0–100 integer
(matching_service stores round(...)); the column was Float. Change it to Integer
for honest typing.

- app/models/property.py: match_score column Float -> Integer.
- Alembic revision c3a4d5e6f7a8 with postgresql_using to cast existing values
  (double precision -> integer isn't implicit). Verified on Postgres 16: an
  existing 87.0 becomes 87, upgrade->downgrade->upgrade is reversible, and
  alembic check reports no drift.
- BDD step now assigns int(score) rather than float(score).

Suite: 150 pytest (95% cov), 9 features / 55 scenarios behave, ruff clean.

+++ backend/features/steps/comparison_steps.py
@@ -109,7 +109,7 @@ def step_property_neighbors(context):
 @given('a property "{name}" has match score {score:d}%')
 def step_property_match_score(context, name, score):
     prop = _get(context, name)
-    prop.match_score = float(score)
+    prop.match_score = int(score)
     context.session.flush()
     context.current_property = prop
